requirements.txt
openai
PyMuPDF
Pillow
google-api-python-client
google-auth